[Fix] Split parameter for port depended and system depended.
Split parameter for port depended and system depended.
Port depended parameter is defined in user application.
System depende parameter is defined in device init code.

Appendix : add EPTPC master and slave sample.

+++ asp/driver/rx_gcc/EPTPC.h
@@ -1,7 +1,52 @@
 /*
- * Copyright (c) 2017- Hisashi Hata
- * Released under the toppers license
- * https://www.toppers.jp/license.html
+ *  TOPPERS/ASP Kernel
+ *      Toyohashi Open Platform for Embedded Real-Time Systems/
+ *      Advanced Standard Profile Kernel
+ *
+ *  Copyright (C) 2017- by Hisashi Hata, JAPAN
+ *
+ * The above copyright holders grant permission gratis to use,
+ * duplicate, modify, or redistribute (hereafter called use) this
+ * software (including the one made by modifying this software),
+ * provided that the following four conditions (1) through (4) are
+ * satisfied.
+ *
+ * (1) When this software is used in the form of source code, the above
+ *     copyright notice, this use conditions, and the disclaimer shown
+ *     below must be retained in the source code without modification.
+ *
+ * (2) When this software is redistributed in the forms usable for the
+ *     development of other software, such as in library form, the above
+ *     copyright notice, this use conditions, and the disclaimer shown
+ *     below must be shown without modification in the document provided
+ *     with the redistributed software, such as the user manual.
+ *
+ * (3) When this software is redistributed in the forms unusable for the
+ *     development of other software, such as the case when the software
+ *     is embedded in a piece of equipment, either of the following two
+ *     conditions must be satisfied:
+ *
+ *   (a) The above copyright notice, this use conditions, and the
+ *       disclaimer shown below must be shown without modification in
+ *       the document provided with the redistributed software, such as
+ *     the user manual.
+ *
+ *   (b) How the software is to be redistributed must be reported to the
+ *       TOPPERS Project according to the procedure described
+ *       separately.
+ *
+ * (4) The above copyright holders and the TOPPERS Project are exempt
+ *     from responsibility for any type of damage directly or indirectly
+ *     caused from the use of this software and are indemnified by any
+ *     users or end users of this software from any and all causes of
+ *     action whatsoever.
+ *
+ * THIS SOFTWARE IS PROVIDED "AS IS." THE ABOVE COPYRIGHT HOLDERS AND
+ * THE TOPPERS PROJECT DISCLAIM ANY EXPRESS OR IMPLIED WARRANTIES,
+ * INCLUDING, BUT NOT LIMITED TO, ITS APPLICABILITY TO A PARTICULAR
+ * PURPOSE. IN NO EVENT SHALL THE ABOVE COPYRIGHT HOLDERS AND THE
+ * TOPPERS PROJECT BE LIABLE FOR ANY TYPE OF DAMAGE DIRECTLY OR
+ * INDIRECTLY CAUSED FROM THE USE OF THIS SOFTWARE.
  */
 
 
@@ -49,7 +94,7 @@ typedef struct PTPCLK_T{
 #define EPTPC_MODE_PORT0 0x10
 #define EPTPC_MODE_PORT1 0x20
 //target depend function
-void eptpc_target_config(uint8_t* mac_ad, uint32_t ipclk_conf, uint32_t stca_sel, uint8_t port_mode);
+void eptpc_target_config(uint8_t* mac_ad, uint32_t ipclk_conf, uint32_t stca_sel, uint32_t port_mode);
 void eptpc_init(PTPINI* pini);
 ER ptp_sync_start(PTPINI* pini);
